scripts/requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.1.0
coverage==7.2.7
idna==3.7
Jinja2==3.1.3
lark==1.1.5
MarkupSafe==2.1.2
python-hcl2==4.3.1
python-terraform==0.10.1
PyYAML==6.0
requests==2.31.0
soupsieve==2.4.1
urllib3==2.0.7
